moved database SessionLocal from utils to core
