+++ api/dependencies/database.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from typing import AsyncGenerator
 
-from api.utils.database import SessionLocal
+from api.core.database import SessionLocal
 
 
 async def get_session() -> AsyncGenerator:
